Update group_vars_mash_servers: remove `default(false)` from the condtion to decide if `mariadb_container_network` needs to be enabled
If either Postgres or MariaDB can be selected for a service, it is sensible to add `default(false)` as running `just optimize` does not remove the configuration, possibly causing an error related to MariaDB (not enabled by default). However, if services which specifically require MariaDB are enabled, then it is reasonable to presume that MariaDB is enabled as well, therefore setting `default(false)` to them is not really expected.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>